UPSTREAM: fsp_broadwell_de: Switch CPU to high frequency mode
According to Yang York the FSP is responsible for switching the CPU into
high frequency mode (HFM). For an unknown reason this is not done for the
BSP on my platform though the APs are switched properly.
This code switches the CPU into HFM which makes sure that all cores are in
high frequency mode before payload is started.

It should not harm the operation even if FSP was successful in switching
to HFM.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia2e05152d0bfa7280d039c66c18eb5e38763c082
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: cf6392f756bac2ef03ff07b7fd5e5d07a76e038d
Original-Change-Id: I91baf538511747d1692a8b6b359df5c3a8d56848
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/19537
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/497397