UPSTREAM: drivers/i2c: Add new driver for RTC type RX6110 SA
This driver enables the usage of the external RTC chip RX6110 SA
(http://www5.epsondevice.com/en/products/i2c/rx6110sab.html) which is
connected to the I2C bus. The I2C address of this device is fixed. One
can change parameters in the device tree so that the used setup can be
adapted to match the configuration of the device on the mainboard.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ied233d7e24c60e475f4866df9fc5d2f0ade239e2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7fd20beaf81455cc29271283426e25e722b90fe1
Original-Change-Id: I1290a10c2d5ad76a317c99c8b92a013309a605d6
Original-Signed-off-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/19625
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/506183
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>