UPSTREAM: nb/intel/gm45: Fix raminit with mixed raw card types
`cardF[n]` should indicate whether the DIMM in channel n is of
raw card type F. However, `cardF[1]` was initialised with the
value meant for `cardF[0]`. This patch results in the correct
initialisation of `cardF`.

Tested on a Lenovo T400 containing two DIMMs: one of raw card
type F and the other of raw card type B. Before the patch, the
system would not boot. After the patch, the system boots with all
of the memory functional.

BUG=none
BRANCH=none
TEST=none

Change-Id: If1d8dd58d2a9953a2d6318efcfecd8ce6d8a5425
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 12e6562289b85a9988048fc731be92966110c622
Original-Change-Id: I7409df0b8c67d7efbdadae39dc718c8df7a92552
Original-Signed-off-by: Tristan Corrick <tristancorrick86@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/19652
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/506210
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>