UPSTREAM: cpu/x86/mtrr: don't guard function declarations
set_var_mtrr() and get_free_var_mtrr() don't need to be guarded
against various stages. It just complicates code which lives
in a compilation unit that is compiled for multiple stages by
needing to reflect the same guarding. Instead, just drop the
declaration guard. earlymtrr.c is still just compiled for earlier
stages, but if needed it's easy to move to a mtrr_util.c that
is compiled for all stages.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2e78ef748b721b2a7ed08250ed0ffcda4dbffa08
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: d86e0e6638062b5d80f5d438f0741dd735734ad4
Original-Change-Id: Id6be6f613771380d5ce803eacf1a0c8b230790b6
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/20018
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Philippe Mathieu-Daud <philippe.mathieu.daude@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/524598